// src/app/interfaces/files.interface.ts
export interface FilesData {
    resultado?: string;
    bucUsuario?: string;
    comentario?: string;
    descRechazo?: string;
    descSubRechazo?: string;
    descripcionTipoDocumento?: string;
    docBase64?: string;
    estado?: string;
    fecha?: string;
    folio?: string;
    id?: string;
    idFilenet?: string;
    idSolicitud?: string;
    idTipoDocumento?: string;
    idTipoRechazo?: string;
    idTipoSubRechazo?: string;
    nombreDocumento?: string;
}